+++ src/main/java/de/fernunihagen/dna/scalephant/network/packages/request/QueryBoundingBoxTimeRequest.java
@@ -1,0 +1,192 @@
+/*******************************************************************************
+ *
+ *    Copyright (C) 2015-2016
+ *  
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *  
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License. 
+ *    
+ *******************************************************************************/
+package de.fernunihagen.dna.scalephant.network.packages.request;
+
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.ByteBuffer;
+
+import de.fernunihagen.dna.scalephant.Const;
+import de.fernunihagen.dna.scalephant.network.NetworkConst;
+import de.fernunihagen.dna.scalephant.network.NetworkPackageDecoder;
+import de.fernunihagen.dna.scalephant.network.NetworkPackageEncoder;
+import de.fernunihagen.dna.scalephant.network.packages.NetworkQueryRequestPackage;
+import de.fernunihagen.dna.scalephant.network.packages.PackageEncodeError;
+import de.fernunihagen.dna.scalephant.network.routing.RoutingHeader;
+import de.fernunihagen.dna.scalephant.storage.entity.BoundingBox;
+import de.fernunihagen.dna.scalephant.storage.entity.SSTableName;
+
+public class QueryBoundingBoxTimeRequest implements NetworkQueryRequestPackage {
+	
+	/**
+	 * The name of the table
+	 */
+	protected final SSTableName table;
+
+	/**
+	 * The the query bounding box
+	 */
+	protected final BoundingBox box;
+	
+	/**
+	 * The timestamp
+	 */
+	protected final long timestamp;
+	
+	/**
+	 * Paging enables
+	 */
+	protected final boolean pagingEnabled;
+	
+	/**
+	 * The max tuples per page
+	 */
+	protected final short tuplesPerPage;
+
+	public QueryBoundingBoxTimeRequest(final String table, final BoundingBox box, final long timestamp, final boolean pagingEnabled, final short tuplesPerPage) {
+		this.table = new SSTableName(table);
+		this.box = box;
+		this.timestamp = timestamp;
+		this.pagingEnabled = pagingEnabled;
+		this.tuplesPerPage = tuplesPerPage;
+	}
+
+	@Override
+	public void writeToOutputStream(final short sequenceNumber, final OutputStream outputStream) throws PackageEncodeError {
+
+		try {
+			final byte[] tableBytes = table.getFullnameBytes();
+			final byte[] bboxBytes = box.toByteArray();
+			
+			final ByteBuffer bb = ByteBuffer.allocate(20);
+			bb.order(Const.APPLICATION_BYTE_ORDER);
+			
+			bb.put(getQueryType());
+			
+			if(pagingEnabled) {
+				bb.put((byte) 1);
+			} else {
+				bb.put((byte) 0);
+			}
+			
+			bb.putShort(tuplesPerPage);
+			
+			bb.putShort((short) tableBytes.length);
+			bb.put(NetworkConst.UNUSED_BYTE);
+			bb.put(NetworkConst.UNUSED_BYTE);
+			bb.putInt((int) bboxBytes.length);
+			bb.putLong(timestamp);
+			
+			// Body length
+			final long bodyLength = bb.capacity() + tableBytes.length + bboxBytes.length;
+			
+			// Unrouted package
+			final RoutingHeader routingHeader = new RoutingHeader(false);
+			NetworkPackageEncoder.appendRequestPackageHeader(sequenceNumber, bodyLength, routingHeader, 
+					getPackageType(), outputStream);
+
+			// Write body
+			outputStream.write(bb.array());
+			outputStream.write(tableBytes);
+			outputStream.write(bboxBytes);
+		} catch (IOException e) {
+			throw new PackageEncodeError("Got exception while converting package into bytes", e);
+		}	
+	}
+	
+	/**
+	 * Decode the encoded package into a object
+	 * 
+	 * @param encodedPackage
+	 * @return
+	 * @throws PackageEncodeError 
+	 */
+	public static QueryBoundingBoxTimeRequest decodeTuple(final ByteBuffer encodedPackage) throws PackageEncodeError {
+		final boolean decodeResult = NetworkPackageDecoder.validateRequestPackageHeader(encodedPackage, NetworkConst.REQUEST_TYPE_QUERY);
+		
+		if(decodeResult == false) {
+			throw new PackageEncodeError("Unable to decode package");
+		}
+		
+	    final byte queryType = encodedPackage.get();
+	    
+	    if(queryType != NetworkConst.REQUEST_QUERY_BBOX_AND_TIME) {
+	    	throw new PackageEncodeError("Wrong query type: " + queryType + " required type is: " + NetworkConst.REQUEST_QUERY_BBOX_AND_TIME);
+	    }
+	    
+	    boolean pagingEnabled = false;
+	    if(encodedPackage.get() != 0) {
+	    	pagingEnabled = true;
+	    }
+	    
+	    final short tuplesPerPage = encodedPackage.getShort();
+
+		final short tableLength = encodedPackage.getShort();
+		
+	    // 2 unused bytes
+	    encodedPackage.get();
+	    encodedPackage.get();
+		
+	    final int bboxLength = encodedPackage.getInt();
+	    final long timestamp = encodedPackage.getLong();
+
+		final byte[] tableBytes = new byte[tableLength];
+		encodedPackage.get(tableBytes, 0, tableBytes.length);
+		final String table = new String(tableBytes);
+		
+		final byte[] bboxBytes = new byte[bboxLength];
+		encodedPackage.get(bboxBytes, 0, bboxBytes.length);
+		final BoundingBox boundingBox = BoundingBox.fromByteArray(bboxBytes);
+		
+		if(encodedPackage.remaining() != 0) {
+			throw new PackageEncodeError("Some bytes are left after decoding: " + encodedPackage.remaining());
+		}
+		
+		return new QueryBoundingBoxTimeRequest(table, boundingBox, timestamp, pagingEnabled, tuplesPerPage);
+	}
+
+	@Override
+	public byte getPackageType() {
+		return NetworkConst.REQUEST_TYPE_QUERY;
+	}
+
+	@Override
+	public byte getQueryType() {
+		return NetworkConst.REQUEST_QUERY_BBOX_AND_TIME;
+	}
+	
+	public SSTableName getTable() {
+		return table;
+	}
+
+	public BoundingBox getBoundingBox() {
+		return box;
+	}
+	
+	public short getTuplesPerPage() {
+		return tuplesPerPage;
+	}
+
+	public boolean isPagingEnabled() {
+		return pagingEnabled;
+	}
+
+	public long getTimestamp() {
+		return timestamp;
+	}
+}